Add fill entries based on InfluxQl time comparisons.
This will add fill entries based on greater than, greater than or
equal to, less than, and less than or equal to comparisons in an
InfluxQl query. So if you specify "WHERE time > 10s AND time < 20s"
in your query, this will add fills between 10s and 20s when it
needs to.